new command

+++ deploy/setup.sh
@@ -40,7 +40,9 @@ supervisorctl restart coconut_api
 
 # Setup Nginx to make your application accessible
 cp $PROJECT_BASE_PATH/deploy/nginx_coconut_api.conf /etc/nginx/sites-available/coconut_api.conf
-rm /etc/nginx/sites-enabled/default
+if [ -f /etc/nginx/sites-enabled/default ]; then
+    rm /etc/nginx/sites-enabled/default
+fi
 ln -s /etc/nginx/sites-available/coconut_api.conf /etc/nginx/sites-enabled/coconut_api.conf
 systemctl restart nginx.service
